fix: bugs beheben & Code-Qualität verbessern

+++ app/controllers/TodoController.php
@@ -27,10 +27,11 @@ class TodoController
     public function store(): void
     {
         $data = json_decode(file_get_contents('php://input'), true);//php:// Stream Wrappers
-        
+
 
         if (empty($data['title'])) {
             Response::json(['error' => 'Title ist Pflicht'], 422);
+            return; // defensive Programmierung sonst geht die Methode weiter
         }
 
         $id = $this->model->create($data);
@@ -38,8 +39,14 @@ class TodoController
     }
 
     // PUT /todos/{id}
+    // PUT /todos/{id}
     public function update(?string $id): void
     {
+        $todo = $this->model->find((int) $id);
+        if (!$todo) {
+            Response::json(['error' => 'Not found'], 404);
+            return;
+        }
         $data = json_decode(file_get_contents('php://input'), true);
         $this->model->update((int) $id, $data);
         Response::json(['message' => 'Aktualisiert']);
@@ -48,6 +55,11 @@ class TodoController
     // DELETE /todos/{id}
     public function destroy(?string $id): void
     {
+        $todo = $this->model->find((int) $id);
+        if (!$todo) {
+            Response::json(['error' => 'Not found'], 404);
+            return;
+        }
         $this->model->delete((int) $id);
         Response::json(['message' => 'Gelöscht (Soft Delete)']);
     }
